review: move logic into use-case
