+++ frontend/src/features/Dashboard/useCases/createDashboard.ts
@@ -1,7 +1,7 @@
 import { dashboardsAPI } from '@api/dashboardsAPI'
 import { addMainWindowBanner } from '@features/MainWindow/useCases/addMainWindowBanner'
 import { sideWindowActions } from '@features/SideWindow/slice'
-import { Level } from '@mtes-mct/monitor-ui'
+import { customDayjs, Level } from '@mtes-mct/monitor-ui'
 import { sideWindowPaths } from 'domain/entities/sideWindow'
 import { generatePath } from 'react-router'
 
@@ -20,7 +20,20 @@ export const createDashboard =
     if (data) {
       dispatch(closeDrawDashboard())
       const newId = `new-${Object.keys(getState().dashboard.dashboards).length}`
-      dispatch(dashboardActions.createDashboard({ extractedArea: data, geom: geometry, id: newId }))
+      const date = customDayjs().format('DD/MM/YYYY')
+      const newDashboardName = `Tab ${date}`
+      const dashboard = {
+        amps: [],
+        controlUnits: [],
+        geom: geometry,
+        id: newId,
+        inseeCode: data.inseeCode,
+        name: newDashboardName,
+        regulatoryAreas: [],
+        reportings: [],
+        vigilanceAreas: []
+      }
+      dispatch(dashboardActions.createDashboard({ dashboard, extractedArea: data }))
       dispatch(sideWindowActions.focusAndGoTo(generatePath(sideWindowPaths.DASHBOARD, { id: newId })))
     }
     if (error) {
